[kernel XHCI]: refactor xhci::rings
