+++ crates/kernel/src/drivers/xhci/rings/mod.rs
@@ -1,0 +1,3 @@
+pub mod command;
+pub mod event;
+pub mod trbs;
